requirements.txt
Django==3.2.13
Pillow==9.1.1
django-admin-sortable2==2.0.4
django-tinymce==3.4.0
environs==9.5.0